wip: fix just a couple, will require way more
