+++ packages/browser/src/components/book/index.ts
@@ -1,2 +1,3 @@
 export { default as BookURLFilterContainer } from '../filters/URLFilterContainer'
 export { BookTable } from './table'
+export { useBookOverview, usePrefetchBook } from './useBookOverview'
