+++ packages/browser/src/components/series/SeriesCard.tsx
@@ -3,7 +3,8 @@ import { Text } from '@stump/components'
 import { FileStatus } from '@stump/graphql'
 import { useCallback } from 'react'
 
-import { usePrefetchSeries, usePrefetchSeriesBooks } from '@/scenes/series'
+import { usePrefetchSeries } from '@/scenes/series'
+import { usePrefetchSeriesBooks } from '@/scenes/series/tabs/books/SeriesBooksScene'
 
 import paths from '../../paths'
 import pluralizeStat from '../../utils/pluralize'
