+++ packages/browser/src/scenes/book/index.ts
@@ -1,3 +1,2 @@
-export { usePrefetchBook } from './BookOverviewScene'
 export { default as BookRouter } from './BookRouter'
 export { usePrefetchBooksAfterCursor } from './BooksAfterCursor'
