+++ packages/browser/src/scenes/library/tabs/books/LibraryBooksScene.tsx
@@ -20,6 +20,7 @@ import {
 import { FilterInput } from '@/components/filters/context'
 import {
 	DEFAULT_MEDIA_ORDER_BY,
+	useMediaURLOrderBy,
 	useSearchMediaFilter,
 	useURLKeywordSearch,
 	useURLPageParams,
@@ -30,7 +31,6 @@ import { EntityTableColumnConfiguration } from '@/components/table'
 import TableOrGridLayout from '@/components/TableOrGridLayout'
 import useIsInView from '@/hooks/useIsInView'
 import { usePreferences } from '@/hooks/usePreferences'
-import { useMediaURLOrderBy } from '@/scenes/bookSearch/BookSearchScene'
 import { useBooksLayout } from '@/stores/layout'
 
 import { useLibraryContext } from '../../context'
