+++ packages/browser/src/scenes/series/index.ts
@@ -1,4 +1,3 @@
 export { useSeriesContext, useSeriesContextSafe } from './context'
 export { usePrefetchSeries } from './SeriesLayout'
 export { default as SeriesRouter } from './SeriesRouter'
-export { usePrefetchSeriesBooks } from './tabs/books/SeriesBooksScene'
